+++ qatrack/units/forms.py
@@ -1,27 +1,20 @@
+from itertools import groupby
 
 from django import forms
-from django.core.exceptions import ValidationError
 from django.conf import settings
+from django.core.exceptions import ValidationError
 from django.utils import timezone
 from django.utils.timezone import timedelta
 from django.utils.translation import ugettext as _
 
-from qatrack.parts import models as p_models
-from qatrack.service_log import models as sl_models
-from qatrack.units import models as u_models
 from qatrack.service_log.forms import HoursMinDurationField
+from qatrack.units import models as u_models
 
 
 def max_24hr(value):
     if value > timedelta(hours=24):
 
-        seconds = value.total_seconds()
-        hours = seconds // 3600
-        mins = (seconds % 3600) // 60
-
-        raise ValidationError(
-            _('Duration can not be greater than 24 hours')
-        )
+        raise ValidationError(_('Duration can not be greater than 24 hours'))
 
 
 year_select = forms.ChoiceField(
@@ -31,23 +24,25 @@ year_select = forms.ChoiceField(
 ).widget.render('year_select', timezone.now().year, attrs={'id': 'id_year_select'})
 
 month_select = forms.ChoiceField(
-        required=False,
-        choices=[
-            (0, 'January'),
-            (1, 'February'),
-            (2, 'March'),
-            (3, 'April'),
-            (4, 'May'),
-            (5, 'June'),
-            (6, 'July'),
-            (7, 'August'),
-            (8, 'September'),
-            (9, 'October'),
-            (10, 'November'),
-            (11, 'December'),
-        ],
-        initial=timezone.now().month - 1
-    ).widget.render('month_select', timezone.now().month - 1, attrs={'id': 'id_month_select'})
+    required=False,
+    choices=[
+        (0, 'January'),
+        (1, 'February'),
+        (2, 'March'),
+        (3, 'April'),
+        (4, 'May'),
+        (5, 'June'),
+        (6, 'July'),
+        (7, 'August'),
+        (8, 'September'),
+        (9, 'October'),
+        (10, 'November'),
+        (11, 'December'),
+    ],
+    initial=timezone.now().month - 1
+).widget.render(
+    'month_select', timezone.now().month - 1, attrs={'id': 'id_month_select'}
+)
 
 
 class UnitAvailableTimeForm(forms.ModelForm):
@@ -134,3 +129,21 @@ class UnitAvailableTimeEditForm(forms.ModelForm):
         if cleaned < self.instance.unit.date_acceptance:
             raise ValidationError('Unit cannot have available time edit before it\'s date of acceptance.')
         return cleaned
+
+
+def unit_site_unit_type_choices(include_empty=False):
+    """Return units grouped by site and unit type, suitable for using as optgroups for select inputs"""
+
+    def site_unit_type(u):
+        return "%s :: %s" % (u.site.name if u.site else "Other", u.type.name)
+
+    def site_unit_name(u):
+        return "%s :: %s" % (u.site.name if u.site else "Other", u.name)
+
+    units = u_models.Unit.objects.select_related("site", "type").order_by("site__name", "type__name", "name")
+    choices = [(ut, list(us)) for (ut, us) in groupby(units, key=site_unit_type)]
+    choices = [(ut, [(u.id, site_unit_name(u)) for u in us]) for (ut, us) in choices]
+    if include_empty:
+        choices = [("", "---------")] + choices
+
+    return choices